Report an error if expected branch not identified

+++ diamond.releng.jenkins/job.scripts/update_single_git_repo_function.sh
@@ -90,6 +90,10 @@ update_single_git_repo_function () {
     # At this point, the repository is clean, and up-to-date with the remote. We need to re-establish the correct local branch.
     # We cannot assume that the local branch is tracking the standard branch, since this might be a Gerrit repository previously used to test a change
     repo_branch=$(grep "${repo_name%.git}" ${WORKSPACE}/artifacts_to_archive/cquery-branches-file.txt | head -n 1 | cut -d "=" -f 2)
+    if [[ -z "${repo_branch}" ]]; then
+        echo "[${repo_name}] Could not identify branch name to switch to, something is wrong. Is this a repo that is no longer in use?"
+        return 1
+    fi
     git -C ${repo_path} checkout --detach --quiet
     git -C ${repo_path} branch -D --quiet ${repo_branch} |& sed "s/^/[${repo_name}] /" || true
     git -C ${repo_path} checkout -b ${repo_branch} remotes/origin/${repo_branch} --no-track --quiet |& sed "s/^/[${repo_name}] /"
